app/requirements.txt
# List your project dependencies here
langchain
langchain-openai
openai
requests
python-dotenv
flask

